logout feature update

+++ src/main/java/com/in28minutes/springboot/web/controller/TodoController.java
@@ -7,6 +7,8 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -21,7 +23,7 @@ import com.in28minutes.springboot.web.model.Todo;
 import com.in28minutes.springboot.web.service.TodoService;
 
 @Controller
-@SessionAttributes("name")
+//@SessionAttributes("name")
 public class TodoController {
 
 	@Autowired
@@ -29,12 +31,12 @@ public class TodoController {
 
 	@InitBinder
 	public void initBinder(WebDataBinder binder) {
-//		Date--> dd/MM/yyyy
+		// Date - dd/MM/yyyy
 		SimpleDateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy");
-		binder.registerCustomEditor(Date.class, new CustomDateEditor(dateFormat, false));
+		binder.registerCustomEditor(Date.class, new CustomDateEditor(
+				dateFormat, false));
 	}
-	
-	
+
 	@RequestMapping(value = "/list-todos", method = RequestMethod.GET)
 	public String showTodos(ModelMap model) {
 		String name = getLoggedInUserName(model);
@@ -42,19 +44,30 @@ public class TodoController {
 		return "list-todos";
 	}
 
-
 	private String getLoggedInUserName(ModelMap model) {
-		return (String) model.get("name");
+		Object principal = SecurityContextHolder.getContext()
+				.getAuthentication().getPrincipal();
+		
+		if (principal instanceof UserDetails) {
+			return ((UserDetails) principal).getUsername();
+		}
+		
+		return principal.toString();
 	}
 
 	@RequestMapping(value = "/add-todo", method = RequestMethod.GET)
 	public String showAddTodoPage(ModelMap model) {
-		model.addAttribute("todo", new Todo(0, getLoggedInUserName(model), "Default Desc", new Date(), false));
+		model.addAttribute("todo", new Todo(0, getLoggedInUserName(model),
+				"Default Desc", new Date(), false));
 		return "todo";
 	}
 
 	@RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam int id) {
+
+		if(id==1)
+			throw new RuntimeException("Something went wrong");
+		
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
@@ -67,7 +80,8 @@ public class TodoController {
 	}
 
 	@RequestMapping(value = "/update-todo", method = RequestMethod.POST)
-	public String updateTodo(ModelMap model, @Valid Todo todo, BindingResult result) {
+	public String updateTodo(ModelMap model, @Valid Todo todo,
+			BindingResult result) {
 
 		if (result.hasErrors()) {
 			return "todo";
